+++ src/cmd/lsmr.cpp
@@ -53,6 +53,6 @@ int main_lsmr(args::Subparser &parser)
   }
   Log::Print(FMT_STRING("All Volumes: {}"), Log::ToNow(all_start));
 
-  WriteOutput(out, coreOpts.iname.Get(), coreOpts.oname.Get(), parser.GetCommand().Name(), coreOpts.keepTrajectory, traj);
+  WriteOutput(coreOpts, out, parser.GetCommand().Name(), traj, Log::Saved());
   return EXIT_SUCCESS;
 }