+++ src/cmd/recon.cpp
@@ -71,7 +71,7 @@ int main_recon(args::Subparser &parser)
       out.chip<4>(iv) = Crop(vol, osz);
     }
     Log::Print(FMT_STRING("All Volumes: {}"), Log::ToNow(all_start));
-    WriteOutput(out, coreOpts.iname.Get(), coreOpts.oname.Get(), parser.GetCommand().Name(), coreOpts.keepTrajectory, traj);
+    WriteOutput(coreOpts, out, parser.GetCommand().Name(), traj, Log::Saved());
   }
 
   return EXIT_SUCCESS;