+++ src/log.cpp
@@ -23,6 +23,7 @@ bool isTTY = false;
 Index progressTarget = -1, progressCurrent = 0, progressNext = 0;
 std::mutex progressMutex;
 std::string progressMessage;
+std::string savedLog;
 } // namespace
 
 Level CurrentLevel()
@@ -45,6 +46,14 @@ void SetLevel(Level const l)
 void SetDebugFile(std::string const &fname)
 {
   debug_file = std::make_shared<HD5::Writer>(fname);
+}
+
+void SaveEntry(std::string const &s) {
+  savedLog.append(s);
+}
+
+auto Saved() -> std::string const &{
+  return savedLog;
 }
 
 void End()